Fix bad type in _Connection class
Since the SocketSubscription is transformed by the `MongoMessageHandler` it is of type `StreamSubscription<MongoReplyMessage>` and not `StreamSubscription<List<int>>`.
The class `_Connection` is private and does not implement any public interface. As such, it is unlikely that fixing the type would break users.

Fwiw, I couldn't find any user of `socketSubscription` in the repository (but I only used github's search, and I am not very familiar with it).

Note: it might be interesting to rename the field, since it's not just a socketSubscription, but a transformed subscription.